+++ docs/other/upload_future_docs.sh
@@ -5,17 +5,19 @@ git checkout v0.7.0
 rm -Rf docs/build/
 cd docs; make html
 cd build
-touch ../../python-future-html-docs.zip
-rm ../../python-future-html-docs.zip
-zip -r ../../python-future-html-docs.zip *
-scp ../../python-future-html-docs.zip python-future.org:
+touch /shared/python-future-html-docs.zip
+rm /shared/python-future-html-docs.zip
+zip -r /shared/python-future-html-docs.zip *
+
+cd /shared
+scp python-future-html-docs.zip python-future.org:
 ssh python-future.org
 
 
 On the remote machine:
 ----------------------
 
-cd /var/www/python-future/html
+cd /var/www/python-future/
 unzip ~/python-future-html-docs.zip
-chmod a+r * _static/*
+chmod a+r * html/* html/_static/*