+++ future/tests/disabled/test_bytes_from_py33.py
@@ -6,7 +6,6 @@ should be modernized).
 """
 from __future__ import (absolute_import, division,
                         print_function, unicode_literals)
-from future import standard_library
 from future.builtins import *
 
 import os