+++ future/tests/test_standard_library.py
@@ -12,7 +12,6 @@ import tempfile
 import os
 import copy
 
-from future.standard_library import RENAMES, REPLACED_MODULES
 from future.tests.base import CodeHandler
 
 
@@ -21,6 +20,7 @@ class TestStandardLibraryRenames(CodeHandler):
     def setUp(self):
         self.interpreter = 'python'
         self.tempdir = tempfile.mkdtemp() + os.path.sep
+        standard_library.install_hooks()
 
     @unittest.skipIf(utils.PY3, 'generic import tests are for Py2 only')
     def test_all(self):
@@ -28,10 +28,10 @@ class TestStandardLibraryRenames(CodeHandler):
         Tests whether all of the old imports in RENAMES are accessible
         under their new names.
         """
-        for (oldname, newname) in RENAMES.items():
+        for (oldname, newname) in standard_library.RENAMES.items():
             if newname == 'winreg' and sys.platform not in ['win32', 'win64']:
                 continue
-            if newname in REPLACED_MODULES:
+            if newname in standard_library.REPLACED_MODULES:
                 # Skip this check for e.g. the stdlib's ``test`` module,
                 # which we have replaced completely.
                 continue
@@ -47,9 +47,9 @@ class TestStandardLibraryRenames(CodeHandler):
         """
         example_PY2_check = False
         with standard_library.suspend_hooks():
-            # An example of code that we don't want to break:
+            # An example of fragile import code that we don't want to break:
             try:
-                import builtins     # fragile check for Python 3.x
+                import builtins
             except ImportError:
                 example_PY2_check = True
         if utils.PY2:
@@ -59,15 +59,50 @@ class TestStandardLibraryRenames(CodeHandler):
         # The import should succeed again now:
         import builtins
 
+    def test_disable_hooks(self):
+        """
+        Tests the old (deprecated) names. These deprecated aliases should be
+        removed by version 1.0
+        """
+        example_PY2_check = False
+
+        standard_library.enable_hooks()   # deprecated name
+        old_meta_path = copy.copy(sys.meta_path)
+
+        standard_library.disable_hooks()
+        if utils.PY2:
+            self.assertTrue(len(old_meta_path) == len(sys.meta_path) + 1)
+        else:
+            self.assertTrue(len(old_meta_path) == len(sys.meta_path))
+
+        # An example of fragile import code that we don't want to break:
+        try:
+            import builtins
+        except ImportError:
+            example_PY2_check = True
+        if utils.PY2:
+            self.assertTrue(example_PY2_check)
+        else:
+            self.assertFalse(example_PY2_check)
+        standard_library.enable_hooks()
+        # The import should succeed again now:
+        import builtins
+        self.assertTrue(len(old_meta_path) == len(sys.meta_path))
+
     def test_remove_hooks(self):
+        """
+        As above, but with the new names
+        """
         example_PY2_check = False
 
         standard_library.install_hooks()
         old_meta_path = copy.copy(sys.meta_path)
-        import builtins
 
         standard_library.remove_hooks()
-        self.assertTrue(len(old_meta_path) == len(sys.meta_path) + 1)
+        if utils.PY2:
+            self.assertTrue(len(old_meta_path) == len(sys.meta_path) + 1)
+        else:
+            self.assertTrue(len(old_meta_path) == len(sys.meta_path))
 
         # An example of fragile import code that we don't want to break:
         try:
@@ -82,28 +117,6 @@ class TestStandardLibraryRenames(CodeHandler):
         # The import should succeed again now:
         import builtins
         self.assertTrue(len(old_meta_path) == len(sys.meta_path))
-
-    def test_remove_hooks2(self):
-        """
-        This verifies that modules like http.client are no longer accessible after
-        disabling import hooks, even if they have been previously imported.
-        
-        The reason for this test is that Python caches imported modules in sys.modules.
-        """
-        standard_library.remove_hooks()
-        try:
-            from . import verify_remove_hooks_affects_imported_modules
-        except RuntimeError as e:
-            self.fail(e.message)
-        finally:
-            standard_library.install_hooks()
-
-    def test_requests(self):
-        """
-        GitHub issue #19: conflict with ``requests``
-        """
-        # The below should succeed while ``requests`` is installed:
-        from . import verify_requests_is_not_broken
 
     @unittest.skipIf(utils.PY3, 'not testing for old urllib on Py3')
     def test_old_urllib_import(self):
@@ -166,23 +179,23 @@ class TestStandardLibraryRenames(CodeHandler):
         self.assertEqual(list(zip_longest(a, b)),
                          [(1, 2), (2, 4), (None, 6)])
 
-    # def test_import_from_module(self):
-    #     """
-    #     Tests whether e.g. "import socketserver" succeeds in a module
-    #     imported by another module. We do not want it to!
-    #     """
-    #     code1 = '''
-    #             from future import standard_library
-    #             import importme2
-    #             '''
-    #     code2 = '''
-    #             import socketserver
-    #             print('Import succeeded!')
-    #             '''
-    #     self._write_test_script(code1, 'importme1.py')
-    #     self._write_test_script(code2, 'importme2.py')
-    #     output = self._run_test_script('importme1.py')
-    #     print(output)
+    def test_import_from_module(self):
+        """
+        Tests whether e.g. "import socketserver" succeeds in a module
+        imported by another module.
+        """
+        code1 = '''
+                from future import standard_library
+                import importme2
+                '''
+        code2 = '''
+                import socketserver
+                print('Import succeeded!')
+                '''
+        self._write_test_script(code1, 'importme1.py')
+        self._write_test_script(code2, 'importme2.py')
+        output = self._run_test_script('importme1.py')
+        print(output)
 
     def test_configparser(self):
         import configparser
@@ -301,3 +314,4 @@ class TestStandardLibraryRenames(CodeHandler):
 
 if __name__ == '__main__':
     unittest.main()
+    standard_library.remove_hooks()